% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Sample barplot}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Iteration count}

\item{time}{Time suspended between each iteration}
}
\value{
A barplot of the sample
}
\description{
Sample barplot
}
